+++ plugin-build/src/main/groovy/io/sentry/android/gradle/util/GroovyCompat.groovy
@@ -1,3 +1,28 @@
+/*
+ * Adapted from https://github.com/bugsnag/bugsnag-android-gradle-plugin/blob/460a71176a990f19f2b6fe61a06f5f138164aa9d/src/main/groovy/com/bugsnag/android/gradle/GroovyCompat.groovy
+ *
+ * Copyright (c) 2015 Bugsnag
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 package io.sentry.android.gradle.util
 
 import org.gradle.api.Project
@@ -6,9 +31,6 @@ import org.gradle.api.Project
  * Contains functions which exploit Groovy's metaprogramming to provide backwards
  * compatibility for older AGP versions that would be impractical to achieve with
  * Kotlin's static type system.
- *
- * Adapted under MIT license from:
- * https://github.com/bugsnag/bugsnag-android-gradle-plugin/blob/master/src/main/groovy/com/bugsnag/android/gradle/GroovyCompat.groovy
  */
 class GroovyCompat {
