+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/util/FieldUtils.kt
@@ -1,0 +1,37 @@
+/*
+ * Adapted from https://github.com/apache/commons-lang/blob/ebcb39a62fc1e47251eceaf63a4b3d731c5227a0/src/main/java/org/apache/commons/lang3/reflect/FieldUtils.java
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.sentry.android.gradle.instrumentation.util
+
+import java.lang.reflect.Field
+
+/**
+ * Gets all fields of the given class and its parents (if any).
+ */
+internal val Class<*>.allFields: List<Field>
+    get() {
+        val allFields = mutableListOf<Field>()
+        var currentClass: Class<*>? = this
+        while (currentClass != null) {
+            val declaredFields = currentClass.declaredFields
+            allFields += declaredFields
+            currentClass = currentClass.superclass
+        }
+        return allFields
+    }
